# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "I-GIFT is a platform for you to give gifts to your friends and family",
        "title": "I-GIFT",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/token": {
            "post": {
                "description": "Generate a new access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Generate a new access token",
                "parameters": [
                    {
                        "description": "the request body for token generation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.GenerateTokenInput"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "usecase.GenerateTokenInput": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}